Modify constructor parameter list for cache
Before we use address bit width to describe the cache, but it seems
that directly describing the byte size is much more human-readable.

What we do care about a cache is the data block size, the number
of ways and the total size of the cache. For the first one, it
determines the granularity of the data we cached; For the second
one, it determines the degree of tolerance when a HASH collision
occurred; For the last one, it indicates how many resources we can
afford for the cache. Therefore, group depth and index width is of
the least significance and is just derived from other parameters.