Modify total latency calculating schematic
The latency of the ram indicates the miss penalty which is enough
to execute both the writing back and the loading. So we do not need
to gather the latency from ram word by word, let alone byte by byte.

As a result, the cpu execution cycles reduce in two magnitudes.
The original byte-by-byte model only reduce the cycles by 1/5, and
with a big data block, cpu geared with cache is slower than directly
using ram.